requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
spacy>=3.6.0
scispacy>=0.5.0
nltk>=3.8.0
textstat>=0.7.0
bert-score>=0.3.13
sentence-transformers>=2.2.0
openai>=1.0.0
anthropic>=0.3.0
mlx>=0.5.0
mlx-lm>=0.5.0
tqdm>=4.65.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
PyYAML>=6.0.0
requests>=2.28.0
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=0.18.0
sphinx-autodoc-typehints>=1.19.0